[exec-tempgen] Integrate Telescope to fzf the templates
telescope.find_files is used to open a telescope window to the user, allowing them to search their templates in their path.

Also, a custom function is added to <CR> mapping to trigger the actual template generation when user selects a file.
The template generation is quite simple, it just takes the selected file, and the current buffer and executes a cp command.